requirements.txt
ale-py==0.10.2
box2d-py==2.3.5
cloudpickle==3.1.1
Farama-Notifications==0.0.4
gymnasium[atari]==1.1.1
numpy==1.25.0
pygame==2.6.1
swig==4.3.1
typing_extensions==4.13.2
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
imageio==2.31.1
ipython==8.14.0
